Update types.h
ocean64: elf64.c:39: build_elf64_image: Assertion `sizeof(struct phdr64) == 0x38' failed.

+++ types.h
@@ -6,9 +6,9 @@
 typedef int8_t i8;
 typedef int16_t i16;
 typedef int32_t i32;
-typedef int32_t i64;
+typedef int64_t i64;
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
-typedef uint32_t u64;
+typedef uint64_t u64;
 #endif
